feat: agora é possivel adicionar produtos ao carrinho e gerar novos pedidos
add: novos endpoints
add: endpoint para pegar informações do usuario logado
att: documentação

+++ src/main/java/br/com/tech/springtech/domain/model/ProdutoPedido.java
@@ -22,7 +22,7 @@ public class ProdutoPedido {
  @EqualsAndHashCode.Include
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer produtoPedidoId;
+    private Long produtoPedidoId;
 
     @Column(nullable = false)
     private Long quantidade;
